Modify Person & PersonTest; override toString() method.
This was a really cool lesson that focused on developing a test BEFORE actually implementing a method. This design methodology helps ensure that the developed methods and classes will meet the specification defined by the tests; in terms of helping a programmer, it is nice because it forces you to think about the END you want your class/method to achieve before you begin implementing code.