Lesson 7, adding the BookTest / Book classes.
This lesson shows, again, how test-driven development will allow a design-specification to be developed before a class is actually implemented. Here, the testBook() method was devised, and then the actual Book class was implemented. First, add a separate BookTest class in the test folder. Then add the Book Class in the src folder. Eclipse has a lot of nice features here such as Ctrl+1 that allows you to 'quick fix' any compiler errors that are shown...this will basically walk you through including all the necessary getters, setters, and fields that you need in your class that you've already built tests for...nice!