Lesson 8, link Book and Person classes.
This is a really important lesson: Combining the Book and Person classes are what is going to give functionality to whatever program I write. There are some confusing parts though, such as why the modifiers aren't necessary in the Book class (hint, it's something to do with the BookTest class). There are also some important things regarding the naming of the new test testGetPerson. This method tests the ability to check who has a book, and return the name of the person, as well as set the person's name who actually has a book checked out. Realize that the setPerson() method is a method of the Book class, not of the Person class, hence why it can be called by a book. Likewise, since it returns a Person object, you can now call person methods on the returned object: hence the call of b2.getPerson().getName() in the test method.