Lesson 9/10: Construct a new class MyLibrary, that makes use of the Person and Book classes.
This is again, a very important lesson, since I am bringing old classes into new classes...i.e., finally making use of them. Again, the idea is to use a test case at first, to then use as a design for the actual class itself. In this instance (LOL), I have created a testMyLibrary method that creates a new MyLibrary object, with a constructor that tests the name of the object, and also makes use of some new assert statements (i.e. assertTrue()). This will ensure that the objects inside the MyLibrary class are actually the correct 'types' and not objects of another type, which would not be allowed.