Lesson 12, testing the checkOut and checkIn methods.
Again, using the test-driven development methodology (boy I really like this), I create a number of different tests, using the assertTrue, assertFalse, and assertEquals methods, to test against different cases. These cases include instances like trying to check out a book that is already checked out, trying to check in a book that was never checked out, just a basic test of the method checkOut, to see if it even works, and so on. The checkOut and checkIn methods are then created in the MyLibrary class using the 'quick fix' methodology from before.