Lesson 12, creation of checkOut and checkIn methods.
Here is where I've implemented the checkOut and checkIn methods in the MyLibrary class. Here, it's important to notice that these methods return True or False values, which enhances their ability to be tested in the MyLibraryTest class. I am using at least one method that the instructor hasn't shown yet, that I'm not sure will work. This is the 'contains' method to see if the book exists. I think it's a good idea to make sure a book exists before I check it out...though, adimittedly, I'm not sure if this will work off the bat...we'll have to see. The other part that I got wrong initially was I was using an empty string to test the getName() method against - this might have worked, but I decided to just go with the 'null' as recommended by the instructor.