Lesson 12, fixes to MyLibrary Class.
These fixes are important, and different from the ones implemented by the instructor. I chose to use the 'fields' to set the person element of the Book class as opposed to using a method to do it. To me, it looks a little bit cleaner. The other difference between our code, is that I include tests to make sure the book is in the list before it can be checked out, and finally, I use the addBook() method to place it back into the library after it has been checked out. This functionality has not been tested for yet, however...I think I'll do that now.