Lesson 13, tests do not pass.
Ok, here, the test should pass, because the most books someone can checkout is currently set to '1', and yet, there are two separete 'checkOut' operations that are executed against Person 'p1'. Hence, the reason for the test, currently not passing - because I'm looking for a false condition as evidence of success, but it's currently being returned as true, which is a bug in the code.