Lesson 13, set up tests for GetBooksForPerson.
This lesson focused on creating the testing code that will ensure that a person is not allowed to check out more books that is set by their maximumBooks field. Here, all the commits will ensure that the person does not get the book, as well as that the book will still remain in the book list (actually, that's just a good idea; it's not currently implemented in this commit).